+++ radp-components/radp-spring-boot-starters/radp-mybatis-spring-boot-starter/src/main/java/space/x9x/radp/mybatis/spring/boot/autoconfigure/MybatisPlusExtensionAutoConfiguration.java
@@ -108,12 +108,7 @@ public class MybatisPlusExtensionAutoConfiguration {
 	@ConditionalOnProperty(name = MybatisPlusExtensionProperties.SQL_REWRITE_ENABLED, havingValue = "true")
 	@Bean
 	public Interceptor columnAliasRewriteInterceptor(MybatisPlusExtensionProperties properties) {
-		MybatisPlusExtensionProperties.SqlRewrite sql = properties.getSqlRewrite();
-		String created = sql.getCreatedColumnName();
-		String updated = sql.getLastModifiedColumnName();
-		String creator = sql.getCreatorColumnName();
-		String updater = sql.getUpdaterColumnName();
-		return new ColumnAliasRewriteInterceptor(created, updated, creator, updater);
+		return new ColumnAliasRewriteInterceptor(properties.getSqlRewrite());
 	}
 
 }